refactor(message-list): move OpenMessageSideEffect to sideeffect.ui package
